Split binary pointwise operator accuracy tests
