+++ tests/test_bitwise_and_scalar.py
@@ -1,0 +1,53 @@
+import random
+
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+
+
+@pytest.mark.bitwise_and_scalar
+@pytest.mark.parametrize("shape", utils.POINTWISE_SHAPES)
+@pytest.mark.parametrize("dtype", utils.INT_DTYPES + utils.BOOL_TYPES)
+def test_bitwise_and_scalar(shape, dtype):
+    if dtype in utils.BOOL_TYPES:
+        inp1 = torch.randint(0, 2, size=shape, dtype=dtype, device="cpu").to(
+            flag_gems.device
+        )
+        inp2 = bool(random.randint(0, 2))
+    else:
+        inp1 = torch.randint(
+            low=-0x7FFF, high=0x7FFF, size=shape, dtype=dtype, device="cpu"
+        ).to(flag_gems.device)
+        inp2 = 0x00FF
+
+    ref_inp1 = utils.to_reference(inp1)
+
+    ref_out = torch.bitwise_and(ref_inp1, inp2)
+    with flag_gems.use_gems():
+        res_out = torch.bitwise_and(inp1, inp2)
+
+    utils.gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.bitwise_and_scalar_
+@pytest.mark.parametrize("shape", utils.POINTWISE_SHAPES)
+@pytest.mark.parametrize("dtype", utils.INT_DTYPES + utils.BOOL_TYPES)
+def test_bitwise_and_scalar_(shape, dtype):
+    if dtype in utils.BOOL_TYPES:
+        inp1 = torch.randint(0, 2, size=shape, dtype=dtype, device=flag_gems.device)
+        inp2 = bool(random.randint(0, 2))
+    else:
+        inp1 = torch.randint(
+            low=-0x7FFF, high=0x7FFF, size=shape, dtype=dtype, device="cpu"
+        ).to(flag_gems.device)
+        inp2 = 0x00FF
+    ref_inp1 = utils.to_reference(inp1.clone())
+
+    ref_out = ref_inp1.bitwise_and_(inp2)
+    with flag_gems.use_gems():
+        res_out = inp1.bitwise_and_(inp2)
+
+    utils.gems_assert_equal(res_out, ref_out)
